refactor(issues): depure types.go — move Filters/CELFilter to options.go
The domain model (Issue/Severity/Source/Category/Ref) no longer imports
internal/filter; the CEL alias + Filters live in options.go. HTTP/MCP own CEL
compilation; the data types stay a pure leaf.
